% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swlrt.R
\name{swlrt}
\alias{swlrt}
\title{Stratified weighted log-rank test}
\usage{
swlrt(df, trt_colname, time_colname, event_colname, strat_colname = NULL,
  wlr = "lr", s_star = NULL, t_star = NULL, rho = NULL,
  gamma = NULL)
}
\arguments{
\item{df}{A data frame. Assume standard structure for time-to-event data.}

\item{trt_colname}{A character string. The name of the treatment column in \code{df}.}

\item{time_colname}{A character string. The name of the column in \code{df} with the survival times.}

\item{event_colname}{A character string. The name of the column in \code{df} with the event status. Assumes 1 means event; 0 means censored.}

\item{strat_colname}{A character string. The name of the stratifying-variable column in \code{df}. Will be converted into a factor. Care is needed here -- the function doesn't do any checks that there are sufficient numbers in each category.}

\item{wlr}{The type of weighted log-rank test. Either the default "lr" for a standard log-rank test, "mw" for a modestly-weighted log-rank test, or "fh" for the Fleming-Harrington rho-gamma family.}

\item{s_star}{This is a parameter for the "mw"-test. Either \code{s_star} or \code{t_star} must be specified. The weights are defined as w(t) = 1 / max(S(t-), \code{s_star}), where S is the Kaplan-Meier estimate in the pooled data.}

\item{t_star}{This is a parameter for the "mw"-test. Either \code{s_star} or \code{t_star} must be specified. The weights are defined as w(t) = 1 / max(S(t-), S(\code{t_star}), where S is the Kaplan-Meier estimate in the pooled data.}

\item{rho}{Rho parameter in the "fh"-test. The weights are defined as w(t) = S(t-)^\code{rho} (1 - S(t-))^\code{gamma}, where S is the Kaplan-Meier estimate in the pooled data.}

\item{gamma}{Gamma parameter in the "fh"-test. The weights are defined as w(t) = S(t-)^\code{rho} (1 - S(t-))^\code{gamma}, where S is the Kaplan-Meier estimate in the pooled data.}
}
\value{
A list containing the overall test result and a data frame with the outcome of the weighted log-rank test in each strata. There is a column \code{o_minus_e_trt} to indicate which treatment the "obs - exp" refers to.
}
\description{
Stratified weighted log-rank test
}
